week_12/day_3/composition_lab/Composition_lab_solution/src/main/java/vehicles/ElectricCar.java
package vehicles;

import components.ElectricMotor;

public class ElectricCar extends Vehicle {
    ElectricMotor electricMotor;

    public ElectricCar(int capacity, double price, String colour, ElectricMotor electricMotor) {
        super(capacity, price, colour);
        this.electricMotor = electricMotor;
    }

    public ElectricMotor getElectricMotor() {
        return electricMotor;
    }

    public int getRangeFromMotorBattery(){
        return electricMotor.getBatteryRange();
    }
}
